Move classes to correct packages
