+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/client/transformation/ExceptionRecordTransformer.java
@@ -1,7 +1,6 @@
 package uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation;
 
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.TransformationRequestCreator;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.response.SuccessfulTransformationResponse;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.model.internal.ExceptionRecord;
