+++ src/test/java/uk/gov/hmcts/reform/bulkscan/orchestrator/client/CaseUpdateRequestCreatorTest.java
@@ -1,6 +1,7 @@
 package uk.gov.hmcts.reform.bulkscan.orchestrator.client;
 
 import org.junit.jupiter.api.Test;
+import uk.gov.hmcts.reform.bulkscan.orchestrator.client.caseupdate.CaseUpdateRequestCreator;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.client.caseupdate.model.request.CaseUpdateDetails;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.client.caseupdate.model.request.CaseUpdateRequest;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.model.internal.ExceptionRecord;
